Rename ILogger
