+++ Adjust/adjust/src/main/java/com/adjust/sdk/Logger.java
@@ -1,22 +1,81 @@
+//
+//  Logger.java
+//  Adjust
+//
+//  Created by Christian Wellenbrock on 2013-04-18.
+//  Copyright (c) 2013 adjust GmbH. All rights reserved.
+//  See the file MIT-LICENSE for copying permission.
+//
+
 package com.adjust.sdk;
 
 import android.util.Log;
 
-public interface Logger {
-    public void setLogLevel(LogLevel logLevel);
+import java.util.Locale;
 
-    public void setLogLevelString(String logLevelString);
+import static com.adjust.sdk.Constants.LOGTAG;
 
-    public void verbose(String message, Object... parameters);
+public class Logger implements ILogger {
 
-    public void debug(String message, Object... parameters);
+    private LogLevel logLevel;
 
-    public void info(String message, Object... parameters);
+    public Logger() {
+        setLogLevel(LogLevel.INFO);
+    }
 
-    public void warn(String message, Object... parameters);
+    @Override
+    public void setLogLevel(LogLevel logLevel) {
+        this.logLevel = logLevel;
+    }
 
-    public void error(String message, Object... parameters);
+    @Override
+    public void setLogLevelString(String logLevelString) {
+        if (null != logLevelString) {
+            try {
+                setLogLevel(LogLevel.valueOf(logLevelString.toUpperCase(Locale.US)));
+            } catch (IllegalArgumentException iae) {
+                error("Malformed logLevel '%s', falling back to 'info'", logLevelString);
+            }
+        }
+    }
 
-    public void Assert(String message, Object... parameters);
+    @Override
+    public void verbose(String message, Object... parameters) {
+        if (logLevel.androidLogLevel <= Log.VERBOSE) {
+            Log.v(LOGTAG, String.format(message, parameters));
+        }
+    }
 
+    @Override
+    public void debug(String message, Object... parameters) {
+        if (logLevel.androidLogLevel <= Log.DEBUG) {
+            Log.d(LOGTAG, String.format(message, parameters));
+        }
+    }
+
+    @Override
+    public void info(String message, Object... parameters) {
+        if (logLevel.androidLogLevel <= Log.INFO) {
+            Log.i(LOGTAG, String.format(message, parameters));
+        }
+    }
+
+    @Override
+    public void warn(String message, Object... parameters) {
+        if (logLevel.androidLogLevel <= Log.WARN) {
+            Log.w(LOGTAG, String.format(message, parameters));
+        }
+    }
+
+    @Override
+    public void error(String message, Object... parameters) {
+        if (logLevel.androidLogLevel <= Log.ERROR) {
+            Log.e(LOGTAG, String.format(message, parameters));
+        }
+    }
+
+    @Override
+    public void Assert(String message, Object... parameters) {
+        Log.println(Log.ASSERT, LOGTAG, String.format(message, parameters));
+    }
 }
